+++ src/s3_log_extraction/ip_utils/_update_ip_to_region_codes.py
@@ -154,7 +154,7 @@ def _get_region_code_from_ip_address(
         return region_string
     except ipinfo.exceptions.RequestQuotaExceededError:  # pragma: no cover
         warnings.warn(
-            msg="IPInfo API request quota exceeded. Returning 'undetermined' value.",
+            "IPInfo API request quota exceeded. Returning 'undetermined' value.",
             category=RuntimeWarning,
             stacklevel=2,
         )